Move grouping functions out of the configurator
