+++ MDANSE/Src/MDANSE/Framework/Jobs/DynamicIncoherentStructureFactor.py
@@ -19,6 +19,9 @@ import collections
 import numpy as np
 from scipy.signal import correlate
 
+from MDANSE.Framework.AtomGrouping.grouping import (
+    add_grouped_totals,
+)
 from MDANSE.Framework.Jobs.IJob import IJob
 from MDANSE.Mathematics.Arithmetic import assign_weights, get_weights, weighted_sum
 from MDANSE.Mathematics.Signal import get_spectrum
@@ -320,7 +323,7 @@ class DynamicIncoherentStructureFactor(IJob):
         )
         self._outputData["disf/s(q,f)/total"].scaling_factor = fact
 
-        self.configuration["grouping_level"].add_grouped_totals(
+        add_grouped_totals(
             self.trajectory,
             self._outputData,
             "disf/f(q,t)",
@@ -328,7 +331,7 @@ class DynamicIncoherentStructureFactor(IJob):
             axis="disf/axes/q|disf/axes/time",
             units="au",
         )
-        self.configuration["grouping_level"].add_grouped_totals(
+        add_grouped_totals(
             self.trajectory,
             self._outputData,
             "disf/s(q,f)",
@@ -346,7 +349,7 @@ class DynamicIncoherentStructureFactor(IJob):
             )
             self._outputData["disf/s(q,f)/ideal/total"].scaling_factor = fact
 
-            self.configuration["grouping_level"].add_grouped_totals(
+            add_grouped_totals(
                 self.trajectory,
                 self._outputData,
                 "disf/s(q,f)/ideal",
