+++ MDANSE/Src/MDANSE/Framework/Jobs/StaticStructureFactor.py
@@ -19,6 +19,10 @@ from collections.abc import Iterator
 import numpy as np
 import numpy.typing as npt
 
+from MDANSE.Framework.AtomGrouping.grouping import (
+    add_grouped_totals,
+    update_pair_results,
+)
 from MDANSE.Framework.Jobs.DistanceHistogram import DistanceHistogram
 from MDANSE.Mathematics.Arithmetic import assign_weights, get_weights, weighted_sum
 
@@ -261,9 +265,7 @@ class StaticStructureFactor(DistanceHistogram):
                     fact1 * np.sum((r**2) * pdfIntra * sincqr, axis=1) * dr,
                 )
 
-        self.configuration["grouping_level"].update_pair_results(
-            self.trajectory, calc_func, self._outputData
-        )
+        update_pair_results(self.trajectory, calc_func, self._outputData)
 
         selected_weights, all_weights = self.trajectory.get_weights(
             prop=self.configuration["weights"]["property"]
@@ -296,7 +298,7 @@ class StaticStructureFactor(DistanceHistogram):
             self._outputData["ssf/total"].scaling_factor = fact
 
             for i in ("/intra", "/inter", ""):
-                self.configuration["grouping_level"].add_grouped_totals(
+                add_grouped_totals(
                     self.trajectory,
                     self._outputData,
                     f"ssf{i}",
